requirements.txt
astroid==2.0.4
autopep8==1.3.5
certifi==2018.8.24
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.1
django-cors-headers==2.4.0
djangorestframework==3.8.2
flake8==3.5.0
gunicorn==19.9.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==2.1.1
pytz==2018.5
requests==2.19.1
rope==0.11.0
six==1.11.0
slackclient==1.2.1
typed-ast==1.1.0
urllib3==1.23
websocket-client==0.51.0
wrapt==1.10.11
